jao-backend/pyproject.toml
[tool.poetry]
name = "jao_backend"
version = "0.1.0"
description = ""
authors = ["Anthoni Gleeson <anthoni.gleeson@digital.trade.gov.uk>", "Stuart Axon stuart.axon@digital.cabinet-office.gov.uk"]
readme = "README.md"
license = "MIT"

packages = [
    { include = "jao_backend", from = "src" }
]

[tool.poetry.dependencies]
jao_backend_schemas = { path = "../jao-backend-schemas", develop = true }
python = ">=3.12,<4"
govuk-frontend-django = "0.15.1"
django = ">=5.0.1,<5.0.15"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
django-webpack-loader = "^3.1.0"
django-components = "^0.102"
django-redis = "^5.4.0"
crispy-forms-gds = "^0.3.1"
pydantic = "^2.7.4"
httpx = "^0.28.0"
hypercorn = "^0.17.3"
wagtail = "^6.3.1"
django-ninja = "^1.3.0"
django-extensions = "^3.2.3"
djantic2 = "^1.0.5"
honcho = "^2.0.0"
celery = "^5.5.1"
flower = "^2.0.1"
celery-singleton = "^0.3.1"
boto3 = "^1.34.64"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
s3-parse-url = "^0.4.0"
pgvector = "^0.4.1"
django-polymorphic = "^3.1.0"
pytest-django = "^4.11.1"
litellm = "^1.68.0"
psycopg2-binary = "^2.9.10"
pytest-unordered = "^0.6.1"
faker = "^37.1.0"
factory-boy = "^3.3.3"
mssql-django = "^1.5"
django-percentagefield = "^0.1.7"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = {extras = ["jupyter"], version = "^24.3.0"}
nbstripout = "^0.7.1"
pytest = "^8.1.1"
pytest-picked = "^0.5.0"
django-debug-toolbar = "4.2"
jupyter = "^1.1.1"


[tool.poetry.group.jupyter.dependencies]
jupyter = "^1.1.1"

[tool.poetry.extras]
jupyter = ["jupyter"]

[tool.poetry.scripts]
manage = "django.core.management:execute_from_command_line"
hypercorn = "hypercorn.__main__:main"

[tool.isort]
profile = "black"

[tool.djlint]
profile="django"
indent="4"
ignore="H021,H006"
exclude="node_modules,staticfiles,static"
custom_blocks="gds_accordion,gds_accordion_item,gds_breadcrumbs,gds_checkboxes,gds_checkbox_conditional,gds_cookie_banner,gds_cookie_banner_message,gds_error_summary,gds_error_summary_error_list_item,gds_footer,gds_footer_nav,gds_footer_meta,gds_header,gds_header_nav_item,gds_phase_banner,gds_summary_list,gds_summary_list_row,gds_summary_list_row_key,gds_summary_list_row_value,gds_summary_list_row_actions,gds_summary_list_row_actions_item,gds_summary_list_card,gds_summary_list_card_actions,gds_summary_list_card_actions_item,gds_tabs,gds_tabs_tab"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "jao_backend.settings.test"
pythonpath = [
  "src"
]
